refactor(challenges): optimize ChallengeHeader and ChallengeTabs comp…

+++ hooks/useChallenges.ts
@@ -1,53 +1,46 @@
-import axios from "axios";
 import {useCallback} from "react";
 import z from "zod";
 import {ChallengeFormSchema} from "@/utils/zod";
 import {useUser} from "@/hooks/useUser";
-
-const api = axios.create({
-    baseURL: "/api",
-    headers: {
-        "Content-Type": "application/json",
-    },
-});
+import {api} from "@/utils/api";
 
 export function useChallenges(userId?: string) {
-    const {refreshUser} = useUser(userId);
+    const {user, refreshUser} = useUser(userId);
 
     const createChallenge = useCallback(async (challengeData: z.infer<typeof ChallengeFormSchema>) => {
         if (!userId) return;
 
         try {
-            const response = await api.post("/user/challenges", {
-                userId,
-                ...challengeData,
-            });
+            const res = await api.post("/user/challenges", {userId, ...challengeData});
             await refreshUser();
-            return response;
+            return res;
         } catch (error) {
             console.error("Erreur:", error);
             throw new Error("Erreur lors de la création du challenge");
         }
-    }, [userId]);
+    }, [userId, refreshUser]);
 
     const deleteChallenge = useCallback(async (challengeId: string) => {
+        if (!userId) return;
+
         try {
             await api.delete(`/user/challenges`, {
-                data: {
-                    userId,
-                    challengeId,
-                },
+                data: {userId, challengeId},
             });
             await refreshUser();
-            return true;
         } catch (error) {
             console.error("Erreur:", error);
             throw new Error("Erreur lors de la suppression du challenge");
         }
-    }, []);
+    }, [userId, refreshUser]);
+
+    const currentChallenges = user?.ReadingGoal?.filter(goal => new Date(goal.deadline) >= new Date()) || [];
+    const pastChallenges = user?.ReadingGoal?.filter(goal => new Date(goal.deadline) < new Date()) || [];
 
     return {
         createChallenge,
         deleteChallenge,
+        currentChallenges,
+        pastChallenges,
     };
 }
